Extract acceptance test build logic into convention plugin
