+++ buildSrc/src/main/kotlin/worblehat.acceptance-test-module.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+    id("worblehat.java-module")
+}
+
+tasks.test {
+    jvmArgumentProviders += listOf(CucumberOptions(project.layout.buildDirectory.dir("cucumber")), VncRecordingDirectoryProvider(project.layout.buildDirectory.dir("testcontainers")))
+}
