+++ worblehat-acceptancetests/build.gradle.kts
@@ -1,5 +1,5 @@
 plugins {
-    id("worblehat.java-module")
+    id("worblehat.acceptance-test-module")
 }
 
 dependencies {
@@ -21,28 +21,4 @@ dependencies {
 
     testImplementation("com.google.guava:guava")
     testImplementation("org.apache.commons:commons-lang3")
-}
-
-tasks.test {
-    jvmArgumentProviders += listOf(CucumberOptions(project.layout.buildDirectory.dir("cucumber")), VncRecordingDirectoryProvider(project.layout.buildDirectory.dir("testcontainers")))
-}
-
-open class CucumberOptions(@OutputDirectory val outputs: Provider<Directory>) : CommandLineArgumentProvider {
-
-    @Internal
-    val htmlReport = outputs.map { it.file("cucumber.html") }
-
-    @Internal
-    val junitReport = outputs.map { it.file("cucumber.xml") }
-
-    @Internal
-    val jsonReport = outputs.map { it.file("cucumber.json") }
-
-    override fun asArguments() =
-        listOf("-Dcucumber.plugin=pretty,html:${htmlReport.get().asFile},junit:${junitReport.get().asFile},json:${jsonReport.get().asFile}")
-}
-
-open class VncRecordingDirectoryProvider(@OutputDirectory val outputDirectory: Provider<Directory>) : CommandLineArgumentProvider {
-
-    override fun asArguments() = listOf("-Dtestcontainers.vnc.recordingDirectory=${outputDirectory.get().asFile}")
 }
